Fixed some errors and left over debug errors
Stuff that is no longer used is being called and needed which will hopefully get rid of any left over errors of null pointers/pathways/whatever the c# equivilent is

+++ Its-a-Whiskey-Buisness/Assets/Scripts/BackgroundLogic/GenericRoom.cs
@@ -458,9 +458,9 @@ public class GenericRoom : MonoBehaviour
     {
         if (RoomNum == sender.GetComponent<ClickDialObject>().Room_num /*|| RoomNum == sender.GetComponent<GenericRoomManager>().Roomnum*/)
         {
-            if (data is int)
+            if (data is float)
             {
-                Temperature = (int)data;
+                Temperature = (float)data;
                 SetTemperature.Raise(this, Temperature);
                 //Raise new temperature
             }
